src/TorahDlApi/requirements.txt
azure-functions
torah-dl
